Adding React Router to Apparelex App
